Converted player ratings values to float

+++ main/db-service.h
@@ -88,10 +88,10 @@ struct Player_Game_Ratings {
   std::string player_email = "";
   int game_id = -1;
 
-  int game_parameter1_value = 0;
-  int game_parameter2_value = 0;
-  int game_parameter3_value = 0;
-  int game_parameter4_value = 0;
+  float game_parameter1_value = 0;
+  float game_parameter2_value = 0;
+  float game_parameter3_value = 0;
+  float game_parameter4_value = 0;
 
   bool is_valid;
 
@@ -118,19 +118,19 @@ struct Joined_Player_Game_Ratings {
 
   std::string game_parameter1_name = "";
   float game_parameter1_weight = 0;
-  int game_parameter1_value = 0;
+  float game_parameter1_value = 0;
 
   std::string game_parameter2_name = "";
   float game_parameter2_weight = 0;
-  int game_parameter2_value = 0;
+  float game_parameter2_value = 0;
 
   std::string game_parameter3_name = "";
   float game_parameter3_weight = 0;
-  int game_parameter3_value = 0;
+  float game_parameter3_value = 0;
 
   std::string game_parameter4_name = "";
   float game_parameter4_weight = 0;
-  int game_parameter4_value = 0;
+  float game_parameter4_value = 0;
 
   std::string category = "";
   int players_per_team = 0;
